Move GhostParticlePartitioner
